requirements.txt
argparse==1.1
pandas==0.19.1
matplotlib==1.5.3
numpy==1.13.3
scipy==0.18.1
seaborn==0.7.1
pillow==4.0.0
scikit-learn==0.18.1
tinify==1.5.0